+++ R/clonalNetwork.R
@@ -55,7 +55,7 @@
 #' @import ggplot2
 #' @importFrom stringr str_sort
 #' @importFrom igraph graph_from_data_frame V `V<-`
-#' @importFrom dplyr %>% group_by select summarize_all count
+#' @importFrom dplyr %>% group_by select summarize_all count n
 #' @importFrom tidygraph as_tbl_graph activate
 #' @importFrom ggraph ggraph geom_edge_bend  geom_node_point scale_edge_colour_gradientn circle guide_edge_colourbar
 #' @importFrom stats setNames
@@ -155,7 +155,7 @@ clonalNetwork <- function(sc.data,
       group_by(meta[,group.by]) %>%
       na.omit() %>%
       unique() %>%
-      summarise(n = n()) %>%
+      summarise(n = dplyr::n()) %>%
       {setNames(.$n, .$`meta[, group.by]`)} 
     
     #Total clones per group.by
@@ -163,7 +163,7 @@ clonalNetwork <- function(sc.data,
       select(all_of(c(cloneCall, group.by))) %>%
       group_by(meta[,group.by]) %>%
       na.omit() %>%
-      summarise(n = n()) %>%
+      summarise(n = dplyr::n()) %>%
       {setNames(.$n, .$`meta[, group.by]`)} 
     
     edge.list <- NULL